web_framework/requirements.txt
Flask==2.0.1
flask-restful==0.3.9
mysql-connector-python==8.0.18